[add] 기타 util 및 테스트 추가
error handle에 대해선 더 생각해야겠다
