+++ src/utils/errors/customError.js
@@ -1,4 +1,5 @@
-class CustomError extends Error {
+// https://medium.com/@Dongmin_Jang/javascript-%EC%97%90%EB%9F%AC-%EC%B2%98%EB%A6%AC-%EB%B0%A9%EB%B2%95-e6cecca61974
+export class CustomError extends Error {
   constructor(code = 'GENERIC', status = 500, ...params) {
     super(...params);
 
@@ -10,5 +11,3 @@ class CustomError extends Error {
     this.status = status;
   }
 }
-
-module.exports = CustomError;
